--- client/src/Components/Support/SupportTeamHeader/SupportTeamData.js
@@ -1,58 +1,0 @@
-export const SupportTeamData = [
-  {
-    supportName: "Desktop Support",
-    supporters: [
-      {
-        name: "Kenny Merzius",
-      },
-      {
-        name: "John Smith",
-      },
-      {
-        name: "Mike Richardson",
-      },
-    ],
-  },
-  {
-    supportName: "Application Support",
-    supporters: [
-      {
-        name: "Mike Fierer",
-      },
-      {
-        name: "Lorenzo Gill",
-      },
-      {
-        name: "Jonathan Guy",
-      },
-    ],
-  },
-  {
-    supportName: "Mobile Support",
-    supporters: [
-      {
-        name: "Ajay Dionne",
-      },
-      {
-        name: "Mike Gill",
-      },
-    ],
-  },
-  {
-    supportName: "Email and Collaboration",
-    supporters: [
-      {
-        name: "Johnny Cage",
-      },
-      {
-        name: "Phil Smith",
-      },
-      {
-        name: "John Cope",
-      },
-      {
-        name: "Michel LaMarre",
-      },
-    ],
-  },
-];
