--- client/src/Components/Support/SupportTeamHeader/SupportTeamHeader.js
@@ -1,12 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-import {NavLink} from "react-router-dom";
-
-function SupportTeamHeader() {
-  return <SupportBanner></SupportBanner>;
-}
-
-const SupportBanner = styled.div``;
-const SupportLink = styled(NavLink)``;
-
-export default SupportTeamHeader;
