+++ client/src/Components/Support/SupportTickets/SupportClosedTickets.js
@@ -1,10 +1,9 @@
 import React from "react";
 import styled from "styled-components";
-import AdminSideBar from "./AdminSideBar";
-import ListItem from "../ListItem/ListItem";
-import TicketSectionHeader from "../ListItem/TicketSectionHeader";
-import Loading from "../Loading";
-import AccountSideBar from "./AccountSideBar";
+import AdminSideBar from "../SideBars/AdminSideBar";
+import AccountSideBar from "../SideBars/AccountSideBar";
+import TicketItem from "../../ListItems/TicketItem";
+import TicketSectionHeader from "../../SectionHeaders/TicketSectionHeader";
 
 function SupportClosedTickets() {
   const [closedTickets, setClosedTickets] = React.useState([]);
@@ -17,14 +16,17 @@ function SupportClosedTickets() {
   return (
     <AdminPage>
       <TicketDashBoard>
-        <AdminSideBar />
+        <SideBar>
+          <AdminSideBar />
+          <AccountSideBar />
+        </SideBar>
         <NewTicketsDisplay>
           <NewTicketItems>
             {closedTickets ? (
               <TicketHeader>
                 <TicketSectionHeader />
                 {closedTickets.map((ticket) => {
-                  return <ListItem ticket={ticket} />;
+                  return <TicketItem ticket={ticket} />;
                 })}
               </TicketHeader>
             ) : (
@@ -36,14 +38,13 @@ function SupportClosedTickets() {
     </AdminPage>
   );
 }
+const SideBar = styled.div`
+  flex: 1;
+`;
 
 const AdminPage = styled.div`
   display: flex;
   flex-direction: column;
-`;
-
-const TicketBanner = styled.h2`
-  text-align: center;
 `;
 
 const TicketDashBoard = styled.div`
@@ -55,7 +56,7 @@ const TicketHeader = styled.div`
 
 const NewTicketItems = styled.div``;
 const NewTicketsDisplay = styled.div`
-  flex: 4;
+  flex: 5;
 `;
 
 export default SupportClosedTickets;
